fix(h3): the audio metric was wrong, not the audio; plus two failure-hiding bugs
E6 reported `video PSNR 31.69 dB | audio SI-SDR -6.88 dB`. The video number is a
pass and confirms the float32 VAE fix. The audio number was measuring the wrong
thing: the VAE returns whole latent frames -- 122 x 800 = 97,600 samples for a
97,333-sample input -- and `a_rec.flatten()[:wave.numel()]` keeps channel 0 aligned
while shifting channel 1 by the 267-sample difference. SI-SDR is brutally sensitive
to alignment, and that shift alone produced the negative value. Truncating on the
time axis before flattening fixes it. Believing the number would have meant hunting
for a bug in the audio VAE that was never there.

Two others, both of which hide failures rather than cause them. The manifest was
written from a `finally`, so a crashed run left `kept: 0` behind and the next run's
"corpus already present" guard skipped an empty dataset straight through to
training. And `torchrun` off PATH resolved to the system python 3.13 torch, which
cannot see these GPUs; going through the venv interpreter pins the launcher to the
same torch as everything else.

The geometry claim is now confirmed against real VAE weights, not just derived:
`73 frames -> 22 latent frames`, exactly the 17n+5 -> 5n+2 in findings.md.

+++ tasks/minimax_h3_20260812T/code/sample.py
@@ -91,7 +91,13 @@ def decode(video_latents, audio_latents, ckpt: Path, device):
     """Denormalize and decode with the frozen H3 VAEs, returning pixels and waveform."""
     from diffusers import AutoencoderKLMiniMaxH3, AutoencoderKLMiniMaxH3Audio
 
-    vae = AutoencoderKLMiniMaxH3.from_pretrained(ckpt / "vae", dtype=torch.bfloat16).to(device).eval()
+    # float32, not bfloat16. The reference's `encode_vae_condition` feeds
+    # `pixels.to(torch.float32)` and this VAE keeps some biases in fp32 (the same
+    # mixed-precision design as the transformer's `_keep_in_fp32_modules`), so a
+    # bf16 input meets an fp32 bias:
+    #   RuntimeError: Input type (c10::BFloat16) and bias type (float) should be the same
+    # At 20.8 GB in fp32 against 94 GB free, there is nothing to buy by narrowing it.
+    vae = AutoencoderKLMiniMaxH3.from_pretrained(ckpt / "vae", dtype=torch.float32).to(device).eval()
     audio_vae = AutoencoderKLMiniMaxH3Audio.from_pretrained(ckpt / "audio_vae",
                                                             dtype=torch.float32).to(device).eval()
 
@@ -101,7 +107,7 @@ def decode(video_latents, audio_latents, ckpt: Path, device):
     a_std = torch.tensor(audio_vae.config.latents_std, device=device).view(1, 1, -1)
 
     latents = video_latents.to(device) * lat_std + lat_mean
-    pixels = vae.decode(latents.to(torch.bfloat16), return_dict=False)[0].float()
+    pixels = vae.decode(latents, return_dict=False)[0].float()
     # Undo the ImageNet normalization the encoder applied over a [0, 1] base.
     mean = torch.tensor(H.PIXEL_MEAN, device=device).view(1, -1, 1, 1, 1)
     std = torch.tensor(H.PIXEL_STD, device=device).view(1, -1, 1, 1, 1)
